+++ index.js
@@ -49,7 +49,7 @@ function runCommand (command, args, ignoreStdError) {
 function base64Encode (file) {
   // read binary data
   var bitmap = fs.readFileSync(file)
-  // convert binary data to base64 encoded string
+    // convert binary data to base64 encoded string
   return new Buffer(bitmap).toString('base64')
 }
 function compareVersions (installed, required) {
@@ -172,6 +172,7 @@ function buildReport (params) {
   var markdownBody = '# Visual Acceptance Tests\n'
   if (markdownReport.new !== '## New\n') {
     markdownBody += markdownReport.new + '\n'
+      if (matches !== null) {
   }
   if (markdownReport.changed !== '## Changed\n') {
     markdownBody += markdownReport.changed
@@ -215,6 +216,64 @@ function buildReport (params) {
   return response
 }
 
+function buildTeamcityBitbucketReport (params, options, prNumber) {
+  return runCommand('phantomjs', ['vendor/html-to-image.js', 'visual-acceptance-report/report.html']).then(function (params) {
+    console.log('Sending to github')
+    var image = base64Encode('images/output.png').replace('data:image\/\w+;base64,', '')
+
+    function uploadToExpress (url, image, name) {
+      var apiOptions = {
+        'headers': {
+          'Content-Type': 'application/json'
+        },
+        'json': {
+          'name': name,
+          'data': image.replace('data:image\/\w+;base64,', '')
+        }
+
+      }
+      var response = request('POST', url + 'api/upload/image', apiOptions)
+      return response.getBody()
+    }
+    var filename = options.project + '-' + options.repo + '-' + prNumber + '.png'
+    uploadToExpress(options.apiUrl, image, filename)
+    var githubApiPostOptions = {
+      'headers': {
+        'user-agent': 'visual-acceptance',
+        'Authorization': 'Basic ' + new Buffer(options.user + ':' + options.password, 'ascii').toString('base64')
+      },
+      'json': {
+        'text': '![PR ember-cli-visual-acceptance Report](' + options.apiUrl + filename + ')'
+      }
+    }
+    var githubApiGetOptions = {
+      'headers': {
+        'user-agent': 'visual-acceptance',
+        'Authorization': 'Basic ' + new Buffer(options.user + ':' + options.password, 'ascii').toString('base64')
+      }
+    }
+    var url = 'http://' + options.domain + '/rest/api/1.0/projects/' + options.project + '/repos/' + options.repo + '/pull-requests/' + prNumber + '/comments'
+    var urlGet = 'http://' + options.domain + '/rest/api/1.0/projects/' + options.project + '/repos/' + options.repo + '/pull-requests/' + prNumber + '/activities'
+    var response = request('GET', urlGet, githubApiGetOptions)
+    var bodyJSON = JSON.parse(response.getBody().toString())
+    var existingComment = false
+    for (var i = 0; i < bodyJSON.values.length; i++) {
+      if (bodyJSON.values[i].action === 'COMMENTED' && bodyJSON.values[i].comment.text.indexOf('![PR ember-cli-visual-acceptance Report]') > -1) {
+        existingComment = true
+        break
+      }
+    }
+    if (existingComment) {
+      response = {
+        error: 'Comment already exists. Just updating image'
+      }
+      console.log('Comment already exists. Just updating image')
+    } else {
+      response = request('POST', url, githubApiPostOptions)
+    }
+    return response
+  })
+}
 module.exports = {
   name: 'ember-cli-visual-acceptance',
   included: function (app) {
@@ -223,13 +282,10 @@ module.exports = {
       app.import(app.bowerDirectory + '/resemblejs/resemble.js', {
         type: 'test'
       })
-      app.import(path.join('vendor', 'bluebird', 'js', 'browser', 'bluebird.min.js'), {
+      app.import(app.bowerDirectory + '/es6-promise/es6-promise.js', {
         type: 'test'
       })
       app.import(path.join('vendor', 'jquery.min.js'), {
-        type: 'test'
-      })
-      app.import(path.join('vendor', 'html2canvas.js'), {
         type: 'test'
       })
       app.import(path.join('vendor', 'VisualAcceptance.js'), {
@@ -358,6 +414,11 @@ module.exports = {
           type: String,
           default: 'visual-acceptance',
           description: 'The ember-cli-visual-acceptance directory where images are save'
+        }, {
+          name: 'build-report',
+          type: Boolean,
+          default: false,
+          description: 'Wheter or not to build a report'
         }],
         run: function (options, rawArgs) {
           var root = this.project.root
@@ -377,7 +438,11 @@ module.exports = {
           }
 
           deleteFolderRecursive(path.join(root, options.imageDirectory))
-          return runCommand('ember', ['br'])
+          if (options.buildReport) {          return runCommand('ember', ['br'])
+            return runCommand('ember', ['br'])
+          } else {
+            return runCommand('ember', ['test'])
+          }
         }
       },
       'travis-visual-acceptance': {
@@ -416,6 +481,106 @@ module.exports = {
           var travisMessage = res.body
           if (/\#new\-baseline\#/.exec(travisMessage)) {
             console.log('Creating new baseline')
+            return runCommand('ember', ['new-baseline', '--image-directory=' + options.imageDirectory, '--build-report=true']).then(function (params) {
+              if (prNumber === false) {
+                console.log('Git add')
+                return runCommand('git', ['add', options.imageDirectory + '/*']).then(function (params) {
+                  console.log('Git commit')
+                  return runCommand('git', ['commit', '-m', '"Adding new baseline images [ci skip]"']).then(function (params) {
+                    console.log('Git push')
+                    return runCommand('git', ['push', 'origin', 'HEAD:' + options.branch], true)
+                  })
+                })
+              } else if (prNumber !== false && prNumber !== 'false' && process.env.VISUAL_ACCEPTANCE_TOKEN) {
+                return buildReport(params).then(buildReport, function (params) {
+                  return buildReport(params).then(function (params) {
+                    throw new Error('Exit 1')
+                  })
+                })
+              } else if (prNumber !== false && prNumber !== 'false' && process.env.VISUAL_ACCEPTANCE_TOKEN) {
+                return runCommand('ember', ['br']).then(buildReport, function (params) {
+                  return buildReport(params).then(function (params) {
+                    throw new Error('Exit 1')
+                  })
+                })
+              } else if (prNumber === false || prNumber === 'false') {
+                return runCommand('ember', ['test']).then(function (params) {
+                  console.log('Git add')
+                  return runCommand('git', ['add', options.imageDirectory + '/*']).then(function (params) {
+                    console.log('Git commit')
+                    return runCommand('git', ['commit', '-m', '"Adding new baseline images [ci skip]"']).then(function (params) {
+                      console.log('Git push')
+                      return runCommand('git', ['push', 'origin', 'HEAD:' + options.branch], true)
+                    })
+                  })
+                })
+              }
+            })
+          }
+        }
+      },
+      'teamcity-bitbucket-visual-acceptance': {
+        name: 'teamcity-bitbucket-visual-acceptance',
+        aliases: ['tbva'],
+        description: 'Run visual-acceptance based off Bitbucket message',
+        works: 'insideProject',
+        availableOptions: [{
+          name: 'image-directory',
+          type: String,
+          default: 'visual-acceptance',
+          description: 'The ember-cli-visual-acceptance directory where images are save'
+        }, {
+          name: 'branch',
+          type: String,
+          default: 'master',
+          description: 'branch to push to'
+        }, {
+          name: 'user',
+          type: String,
+          default: '',
+          description: 'Bitbucket username'
+        }, {
+          name: 'password',
+          type: String,
+          default: '',
+          description: 'Bitbucket user\'s password'
+        }, {
+          name: 'domain',
+          type: String,
+          default: '',
+          description: 'Domain of Bitbucket server ex("bitbucket.host.com")'
+        }, {
+          name: 'project',
+          type: String,
+          default: '',
+          description: 'Name of project where the repository is held'
+        }, {
+          name: 'repo',
+          type: String,
+          default: '',
+          description: 'Name of the repository'
+        }, {
+          name: 'api-url',
+          type: String,
+          default: '',
+          description: 'Url of api server to save and host images. https://gitlab.com/EWhite613/express-reports'
+        }],
+        run: function (options, rawArgs) {
+          if (options.user.length === 0 || options.password.length === 0 || options.domain.length === 0 || options.project.length === 0 || options.repo.length === 0 || options.apiUrl.length === 0) {
+            console.log('Need to supply a user, password, domain, project, repo, and express-url . Sorry the bitbucket api sucks. \n Just running ember test')
+            return runCommand('ember', ['test'])
+          }
+
+          if (process.env.TEAMCITY_PULL_REQUEST === null) {
+            console.log('No Teamcity found. Just running ember test')
+            return runCommand('ember', ['test'])
+          }
+          var prNumber = process.env.TEAMCITY_PULL_REQUEST
+          var baseUrl = 'http://' + options.domain + '/rest/api/1.0/projects/' + options.project + '/repos/' + options.repo + '/pull-requests/' + prNumber
+          var res = request('GET', baseUrl)
+          var PrDescription = JSON.parse(res.body).description
+          if (PrDescription && /\#new\-baseline\#/.exec(PrDescription)) {
+            console.log('Creating new baseline')
             return runCommand('ember', ['new-baseline', '--image-directory=' + options.imageDirectory]).then(function (params) {
               if (prNumber === false) {
                 console.log('Git add')
@@ -428,13 +593,15 @@ module.exports = {
                 })
               }
             })
-          } else if (prNumber !== false && prNumber !== 'false' && process.env.VISUAL_ACCEPTANCE_TOKEN) {
-            return runCommand('ember', ['br']).then(buildReport, function (params) {
-              return buildReport(params).then(function (params) {
+          } else if (prNumber !== false) {
+            return runCommand('ember', ['br']).then(function (params) {
+              return buildTeamcityBitbucketReport(params, options, prNumber)
+            }, function (params) {
+              return buildTeamcityBitbucketReport(params, options, prNumber).then(function (params) {
                 throw new Error('Exit 1')
               })
             })
-          } else if ((prNumber === false || prNumber === 'false')) {
+          } else if (prNumber === false) {
             return runCommand('ember', ['test']).then(function (params) {
               console.log('Git add')
               return runCommand('git', ['add', options.imageDirectory + '/*']).then(function (params) {