+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+# 0.1.13
+* Add Teamcity/Bitbucket comment on PR integration
+
+# 0.1.12
+No CHANGELOG section found in Pull Request description.
+Use a `# CHANGELOG` section in your Pull Request description to auto-populate the `CHANGELOG.md`
+
+# 0.1.11
+Updating instructions.
+
 # 0.1.10
 Updating instructions.